refactor(Tour): move parseSteps to utils.js
